+++ src/main/java/fr/insalyon/creatis/moteurlite/boutiques/BoutiquesService.java
@@ -2,24 +2,24 @@ package fr.insalyon.creatis.moteurlite.boutiques;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import fr.insalyon.creatis.moteurlite.MoteurLiteException;
 
-/**
- * Author: Sandesh Patil [https://github.com/sandepat]
- */
+import fr.insalyon.creatis.moteurlite.MoteurLiteException;
+import fr.insalyon.creatis.moteurlite.boutiques.scheme.BoutiquesDescriptor;
+import fr.insalyon.creatis.moteurlite.boutiques.scheme.Input;
+import fr.insalyon.creatis.moteurlite.boutiques.scheme.OutputFile;
+
 
 public class BoutiquesService {
 
-    public BoutiquesService() {
-    }
+    public BoutiquesService() {}
 
-    // Method to parse the boutiques descriptor file
     public BoutiquesDescriptor parseFile(String boutiquesDescriptorFile) throws MoteurLiteException {
         try {
             ObjectMapper objectMapper = new ObjectMapper();
@@ -29,31 +29,21 @@ public class BoutiquesService {
         }
     }
 
-    public HashMap<String, Input.Type> getInputTypes(BoutiquesDescriptor boutiquesDescriptor) {
-        HashMap<String, Input.Type> inputTypes = new HashMap<>();
-        for (Input input : boutiquesDescriptor.getInputs()) {
-            inputTypes.put(input.getId(), input.getType());
-        }
-        return inputTypes;
+    public Map<String, Input.Type> getInputTypes(BoutiquesDescriptor boutiquesDescriptor) {
+        return boutiquesDescriptor.getInputs().stream()
+            .collect(Collectors.toMap(Input::getId, Input::getType));
     }
 
-    public HashMap<String, Input> getInputsMap(BoutiquesDescriptor boutiquesDescriptor) {
-        HashMap<String, Input> inputTypes = new HashMap<>();
-        for (Input input : boutiquesDescriptor.getInputs()) {
-            inputTypes.put(input.getId(), input);
-        }
-        return inputTypes;
+    public Map<String, Input> getInputsMap(BoutiquesDescriptor boutiquesDescriptor) {
+        return boutiquesDescriptor.getInputs().stream()
+            .collect(Collectors.toMap(Input::getId, Function.identity()));
     }
 
-    public HashMap<String, OutputFile> getOutputMap(BoutiquesDescriptor boutiquesDescriptor) {
-        HashMap<String, OutputFile> outputFiles = new HashMap<>();
-        for (OutputFile outputFile : boutiquesDescriptor.getOutputFiles()) {
-            outputFiles.put(outputFile.getId(), outputFile);
-        }
-        return outputFiles;
+    public Map<String, OutputFile> getOutputMap(BoutiquesDescriptor boutiquesDescriptor) {
+        return boutiquesDescriptor.getOutputFiles().stream()
+            .collect(Collectors.toMap(OutputFile::getId, Function.identity()));
     }
 
-    // Method to get optional inputs from the BoutiquesDescriptor
     public Set<String> getInputOptionalOfBoutiquesFile(BoutiquesDescriptor boutiquesDescriptor) {
         Set<String> optionalInputs = new HashSet<>();
         for (Input input : boutiquesDescriptor.getInputs()) {
@@ -64,23 +54,21 @@ public class BoutiquesService {
         return optionalInputs;
     }
 
-    // Method to extract "crossSet", "dotSet", and "containerSet" from the custom field of BoutiquesDescriptor
     public Set<String> getCrossMap(BoutiquesDescriptor boutiquesDescriptor) {
-        return extractCustomField(boutiquesDescriptor, "VIP:cross");
+        return extractCustomField(boutiquesDescriptor, "vip:cross");
     }
 
     public Set<String> getDotMap(BoutiquesDescriptor boutiquesDescriptor) {
-        return extractCustomField(boutiquesDescriptor, "VIP:dot-inputs");
+        return extractCustomField(boutiquesDescriptor, "vip:dot");
     }
 
     public Set<String> getContainerSet(BoutiquesDescriptor boutiquesDescriptor) {
-        return extractCustomField(boutiquesDescriptor, "VIP:imagepath");
+        return extractCustomField(boutiquesDescriptor, "vuo:imagepath");
     }
 
-    // Helper method to extract the relevant values from the custom field
     private Set<String> extractCustomField(BoutiquesDescriptor boutiquesDescriptor, String key) {
         Set<String> resultSet = new HashSet<>();
-        Map<String, Object> customMap = boutiquesDescriptor.getCustom().getAdditionalProperties(); // Assuming getAdditionalProperties() exists for custom
+        Map<String, Object> customMap = boutiquesDescriptor.getCustom().getAdditionalProperties();
 
         if (customMap != null && customMap.containsKey(key)) {
             Object value = customMap.get(key);
