+++ src/main/java/fr/insalyon/creatis/moteurlite/boutiques/scheme/ConditionalPathTemplate.java
@@ -1,5 +1,5 @@
 
-package fr.insalyon.creatis.moteurlite.boutiques;
+package fr.insalyon.creatis.moteurlite.boutiques.scheme;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -8,22 +8,30 @@ import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
-
-/**
- * Ids of the inputs that are disabled when the corresponding value choice is selected.
- * 
- */
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-
+    "propertyNames"
 })
 @Generated("jsonschema2pojo")
-public class ValueDisables {
+public class ConditionalPathTemplate {
 
+    @JsonProperty("propertyNames")
+    private Object propertyNames;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new LinkedHashMap<String, Object>();
+
+    @JsonProperty("propertyNames")
+    public Object getPropertyNames() {
+        return propertyNames;
+    }
+
+    @JsonProperty("propertyNames")
+    public void setPropertyNames(Object propertyNames) {
+        this.propertyNames = propertyNames;
+    }
 
     @JsonAnyGetter
     public Map<String, Object> getAdditionalProperties() {
