+++ src/main/java/fr/insalyon/creatis/moteurlite/boutiques/scheme/InvocationSchema.java
@@ -1,5 +1,5 @@
 
-package fr.insalyon.creatis.moteurlite.boutiques;
+package fr.insalyon.creatis.moteurlite.boutiques.scheme;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -8,30 +8,17 @@ import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-    "propertyNames"
+
 })
 @Generated("jsonschema2pojo")
-public class ConditionalPathTemplate {
+public class InvocationSchema {
 
-    @JsonProperty("propertyNames")
-    private Object propertyNames;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new LinkedHashMap<String, Object>();
-
-    @JsonProperty("propertyNames")
-    public Object getPropertyNames() {
-        return propertyNames;
-    }
-
-    @JsonProperty("propertyNames")
-    public void setPropertyNames(Object propertyNames) {
-        this.propertyNames = propertyNames;
-    }
 
     @JsonAnyGetter
     public Map<String, Object> getAdditionalProperties() {
