+++ src/main/java/fr/insalyon/creatis/moteurlite/boutiques/scheme/Tags.java
@@ -1,5 +1,5 @@
 
-package fr.insalyon.creatis.moteurlite.boutiques;
+package fr.insalyon.creatis.moteurlite.boutiques.scheme;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -12,7 +12,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 
 /**
- * Ids of the inputs that are required when the corresponding value choice is selected.
+ * A set of key-value pairs specifying tags describing the pipeline. The tag names are open, they might be more constrained in the future.
  * 
  */
 @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -20,18 +20,18 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 })
 @Generated("jsonschema2pojo")
-public class ValueRequires {
+public class Tags {
 
     @JsonIgnore
-    private Map<String, Object> additionalProperties = new LinkedHashMap<String, Object>();
+    private Map<String, String> additionalProperties = new LinkedHashMap<String, String>();
 
     @JsonAnyGetter
-    public Map<String, Object> getAdditionalProperties() {
+    public Map<String, String> getAdditionalProperties() {
         return this.additionalProperties;
     }
 
     @JsonAnySetter
-    public void setAdditionalProperty(String name, Object value) {
+    public void setAdditionalProperty(String name, String value) {
         this.additionalProperties.put(name, value);
     }
