+++ src/main/java/fr/insalyon/creatis/moteurlite/boutiques/scheme/TestOutputFile.java
@@ -1,5 +1,5 @@
 
-package fr.insalyon.creatis.moteurlite.boutiques;
+package fr.insalyon.creatis.moteurlite.boutiques.scheme;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
