+++ src/main/java/fr/insalyon/creatis/moteurlite/boutiques/scheme/ValueDisables.java
@@ -1,5 +1,5 @@
 
-package fr.insalyon.creatis.moteurlite.boutiques;
+package fr.insalyon.creatis.moteurlite.boutiques.scheme;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -12,7 +12,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 
 /**
- * A set of key-value pairs specifying tags describing the pipeline. The tag names are open, they might be more constrained in the future.
+ * Ids of the inputs that are disabled when the corresponding value choice is selected.
  * 
  */
 @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -20,18 +20,18 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 })
 @Generated("jsonschema2pojo")
-public class Tags {
+public class ValueDisables {
 
     @JsonIgnore
-    private Map<String, String> additionalProperties = new LinkedHashMap<String, String>();
+    private Map<String, Object> additionalProperties = new LinkedHashMap<String, Object>();
 
     @JsonAnyGetter
-    public Map<String, String> getAdditionalProperties() {
+    public Map<String, Object> getAdditionalProperties() {
         return this.additionalProperties;
     }
 
     @JsonAnySetter
-    public void setAdditionalProperty(String name, String value) {
+    public void setAdditionalProperty(String name, Object value) {
         this.additionalProperties.put(name, value);
     }
